requirements.txt
flask==2.0.1
tensorflow==2.10.0
numpy
pillow
matplotlib
scikit-learn
tqdm
werkzeug
